udev: link-config - fix corruption
The parser used for MTU and Speed expects them to be size_t, not unsigned int.

This caused a corruption in the rest of the structure.

Reported by David O Neill <david.m.oneill@intel.com>.

+++ src/udev/net/link-config.h
@@ -66,8 +66,8 @@ struct link_config {
         NamePolicy *name_policy;
         char *name;
         char *alias;
-        unsigned int mtu;
-        unsigned int speed;
+        size_t mtu;
+        size_t speed;
         Duplex duplex;
         WakeOnLan wol;
